+++ refl1d/lib/reflcalc.h
@@ -135,8 +135,8 @@ magnetic_amplitude(const int layers,
                    Cplx Ra[], Cplx Rb[], Cplx Rc[], Cplx Rd[]);
 
 void
-resolution(int Nin, const double Qin[], const double Rin[],
-	   int N, const double Q[], const double dQ[], double R[]);
+resolution(size_t Nin, const double Qin[], const double Rin[],
+	   size_t N, const double Q[], const double dQ[], double R[]);
 
 int
 contract_by_step(int n, double d[], double sigma[],
